requirements.txt
dvc 
dagshub 
pandas 
numpy 
scikit-learn 
joblib 
requests 
pyyaml